requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
openpyxl==3.1.5
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
pytesseract==0.3.13
pywin32==308
